--- packages/web/src/components/PopupModal/index.tsx
@@ -1,1 +1,0 @@
-export * from './PopupModal';
