+++ packages/web/src/pages/index.tsx
@@ -1,9 +1,10 @@
 import { NextPage } from 'next';
 import { Config } from '@hangar/shared';
-import { Flex, Heading, Text } from '@chakra-ui/react';
+import { Flex, Heading, Text, Box } from '@chakra-ui/react';
 import React from 'react';
 import { PageContainer } from '../components/layout/PageContainer';
 import { Prizes } from '../components/Prizes';
+import { RegistrationModal } from '../components/RegistrationModal';
 import { usePrizesStore } from '../stores/prizes';
 
 const Home: NextPage = () => {
@@ -20,6 +21,9 @@ const Home: NextPage = () => {
       heading={Config.homepage.welcome.title}
       isLoading={!prizesFetched}
     >
+      <Box>
+        <RegistrationModal />
+      </Box>
       <Flex direction="column" gap={5}>
         {Config.homepage.welcome.content.map((text) => (
           <Text key={text.substring(0, 15)} fontSize="xl">
